Create ThsAveragercsv.py
This is the main class containing the constructor that reads in the input csv file with data, the name of the output file, the number of samples and the initial and final temperatures. It also contains the main method, Average_Calculate to do the actual averaging and write the averaged data into the output csv file.